pyproject.toml
[tool.poetry]
name = "webtronicstask"
version = "0.1.0"
description = ""
authors = ["onehandedpirate <plagueismkii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.100.0"}
sqlalchemy = "^2.0.18"
alembic = "^1.11.1"
asyncpg = "^0.28.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
